+++ apiserver/metricsender/package_test.go
@@ -1,0 +1,13 @@
+// Copyright 2014 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package metricsender_test
+
+import (
+	"github.com/juju/juju/testing"
+	stdtesting "testing"
+)
+
+func TestPackage(t *stdtesting.T) {
+	testing.MgoTestPackage(t)
+}
