+++ apiserver/metricsmanager/metricsmanager.go
@@ -10,9 +10,9 @@ import (
 	"github.com/juju/loggo"
 
 	"github.com/juju/juju/apiserver/common"
+	"github.com/juju/juju/apiserver/metricsender"
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/state/metricsender"
 )
 
 var (
